% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nblomax.R
\name{llcurve_lomax}
\alias{llcurve_lomax}
\title{Log-log curve for negative binomial-Lomax PMF}
\usage{
llcurve_lomax(xmax, lpar = c(1, 1, 1), lik = c("nb", "geom", "poi"),
  q = 1000, add = TRUE, ...)
}
\arguments{
\item{xmax}{the maximum quantile at which to evaluate the PMF.}

\item{lpar}{numeric vector containing the tail, scale, and shape parameters
see \code{\link{dnblomax}} for details.}

\item{lik}{choice of negative binomial, geometric, or Poisson likelihood}

\item{q}{positive integer number of quadrature points, increase for better
accuracy}

\item{add}{logical, should the curve be added to the existing plot window?}

\item{...}{additional arguments passed to \code{\link[graphics]{curve}}}
}
\value{
a list with components x and y of the points that were drawn is
  returned invisibly.
}
\description{
Draw the probability mass function of a negative binomial-Lomax
distribution with both horizontal and vertical axes log transformed.
}
\seealso{
\code{\link{lpmf}}
}
