% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poilog.R
\name{llcurve_nb}
\alias{llcurve_nb}
\title{Log-log curve for negative binomial PMF}
\usage{
llcurve_nb(xmax, lpar, add = TRUE, ...)
}
\arguments{
\item{xmax}{the maximum quantile at which to evaluate the PMF.}

\item{lpar}{numeric vector containing the parameters size,mu
see \code{\link[stats]{dnbinom}} for details.}

\item{add}{logical, should the curve be added to the existing plot window?}

\item{...}{additional arguments passed to \code{\link[graphics]{curve}}}
}
\value{
a list with components x and y of the points that were drawn is
  returned invisibly.
}
\description{
Draw the probability mass function of a negative binomial
  distribution with both horizontal and vertical axes log transformed.
}
