% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poilog.R
\name{llcurve_poilog}
\alias{llcurve_poilog}
\title{Log-log curve for Poisson-lognormal PMF}
\usage{
llcurve_poilog(xmax, lpar, add = TRUE, use_sads = TRUE, quadpts = 1000, ...)
}
\arguments{
\item{xmax}{the maximum quantile at which to evaluate the PMF.}

\item{lpar}{numeric vector containing the parameters mu,sig
see \code{\link[sads]{dpoilog}} for details.}

\item{add}{logical, should the curve be added to the existing plot window?}

\item{use_sads}{logical, should the sads package be used to compute
\code{\link[sads]{dpoilog}}? If not, an internal custom implementation is used.
The custom implementation is less numerically stable than sads.}

\item{quadpts}{positive integer number of quadrature points, increase for
better accuracy}

\item{...}{additional arguments passed to \code{\link[graphics]{curve}}}
}
\value{
a list with components x and y of the points that were drawn is
  returned invisibly.
}
\description{
Draw the probability mass function of a Poisson-lognormal
distribution with both horizontal and vertical axes log transformed.
}
\seealso{
\code{\link{lpmf}}
}
