% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quminorm.R
\name{lpmf}
\alias{lpmf}
\title{Log-log plot}
\usage{
lpmf(
  x,
  bw = 0.25,
  logbase = NULL,
  discrete = TRUE,
  midpt = FALSE,
  bin_thresh = 10,
  add = FALSE,
  doplot = TRUE,
  ...
)
}
\arguments{
\item{x}{a vector of non-negative integers or decimal values (the data).}

\item{bw}{the bandwidth of the histogram. Increase for more smoothing,
decrease to make the plot capture small-scale variation.}

\item{logbase}{the base of the logarithmic transformation. Defaults to
natural log.}

\item{discrete}{logical; if TRUE, x should contain only integers. In this
case, the unique values of x with large counts are not combined into bins.}

\item{midpt}{logical; if TRUE, the histogram is evaluated at the midpoint of
the logarithmic bins by geometric mean. If FALSE, uses the left endpoint
of the bin.}

\item{bin_thresh}{positive integer; if x is discrete and the mode of the
distribution is not zero, finds the smallest unique x value where the
count is less than bin_thresh. Values larger than this are placed into
bins, while values smaller than this are not binned.}

\item{add}{logical; if TRUE add to an already existing plot using
\code{lines}. Otherwise make a new plot using \code{plot}.}

\item{doplot}{logical; if TRUE produces a plot. Otherwise returns a data
frame containing the horizontal and vertical coordinates of the plot.}

\item{...}{additional arguments passed to \code{lines} or \code{plot}.}
}
\description{
Produces a histogram with logarithmic bins and with both the horizontal
and vertical axes log-transformed. Enables visualization of data from
heavy-tailed distributions.
}
\details{
Heavy-tailed data such as those following power-law type distributions are
difficult to visualize using traditional histograms, because a small number
of points have extremely large values, while all the other points have small
values. A solution is to use logarithmic bins, where the spacing of breaks
increases multiplicatively from left to right. If the x-axis is then
transformed, these bins appear equally spaced. If the y-axis is also
log-transformed, this facilitates comparison to a power-law, since such data
will appear to follow a straight line in a log-log plot. The slope of the
line indicates the exponent of the power law.

Note that while traditional histograms produce barplots, lpmf produces either
points or lines. The reason for this is the log-transformation of the y-axis
creates negative values, making barplots unattractive. Also, using points
and lines enables multiple data vectors' distributions to be overlayed on
a single plot to facilitate comparisons, whereas it is difficult to stack
many separate histograms in a single plot.
}
