% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poilog.R
\name{nb_mle}
\alias{nb_mle}
\title{Negative binomial MLEs}
\usage{
nb_mle(x, ...)
}
\arguments{
\item{x}{vector of non-negative integers (the data).}

\item{...}{additional arguments passed to \code{\link[fitdistrplus]{fitdist}}.}
}
\value{
a named numeric vector containing the MLEs of size,mu (see
  \code{\link[stats]{dnbinom}} for details).
  The maximized value of the log-likelihood is attached as an attribute.
}
\description{
Compute the maximum likelihood estimates for parameters of the
  negative binomial distribution.
}
