% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poilog.R
\name{nb_mle_matrix}
\alias{nb_mle_matrix}
\title{Negative binomial MLEs for columns of a matrix}
\usage{
nb_mle_matrix(m, ...)
}
\arguments{
\item{m}{a matrix or sparse Matrix of non-negative integers (the data).}

\item{...}{additional arguments passed to \code{\link{nb_mle}}.}
}
\value{
a data frame whose rows correspond to the columns of m.
  The columns contain the MLEs for the parameters, the log-likelihood, and
  the \code{\link[stats]{BIC}} values.
}
\description{
Compute the maximum likelihood estimates for parameters of the
  negative binomial distribution for each column of a matrix.
}
