% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poilog.R
\name{nb_pzero2mu}
\alias{nb_pzero2mu}
\title{Estimate negative binomial mean parameter from zero fraction}
\usage{
nb_pzero2mu(lpz, size)
}
\arguments{
\item{lpz}{a numeric vector representing the natural log of the empirical
fraction of zeros from count data.}

\item{size}{either a positive scalar or a vector of same length as lpz
specifying the fixed size parameter (see \code{\link[stats]{dnbinom}}).}
}
\value{
a numeric vector of estimated scale parameters for each element of
  lpz.
}
\description{
Given a fixed size parameter, use the fraction of zeros to
  estimate the mean parameter of a negative binomial
  distribution via the method of moments.
}
