% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nblomax.R
\name{nblomax_mle}
\alias{nblomax_mle}
\title{Negative binomial-Lomax MLEs}
\usage{
nblomax_mle(x, quadpts = 1000, m = rep(0, 3), s = rep(Inf, 3),
  shape = NULL, om = "BFGS", ...)
}
\arguments{
\item{x}{vector of non-negative integers (the data).}

\item{quadpts}{positive integer, number of quadrature points.}

\item{m}{numeric vector of length three. The prior means for the log of the
tail, scale, and shape parameters. See \code{\link{dnblomax}} for details.}

\item{s}{numeric vector of length three. The prior standard deviations for
the log of the tail, scale, and shape parameters. If s is finite, an L2
(ridge) penalty shrinks the MLEs toward m. If s is Inf, the unpenalized MLE
is obtained.}

\item{shape}{positive scalar. If NULL the shape parameter is estimated from
data. If not NULL, the shape parameter is fixed at the specified value
and only tail and scale are estimated. Shape=Inf corresponds to
Poisson-Lomax.}

\item{om}{optimization method to be used by \code{\link[stats]{optim}}.}

\item{...}{additional arguments passed to \code{\link[stats]{optim}}}
}
\value{
a named numeric vector containing the MLEs of tail, scale, and shape
  the maximized value of the log-likelihood is attached as an attribute.
}
\description{
Compute the maximum likelihood estimates for parameters of the
negative binomial-Lomax distribution.
}
