% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nblomax.R
\name{nblomax_mle_matrix}
\alias{nblomax_mle_matrix}
\title{Negative binomial-Lomax MLEs for columns of a matrix}
\usage{
nblomax_mle_matrix(m, quadpts = 1000, maxtry = 10, verbose = FALSE, ...)
}
\arguments{
\item{m}{a matrix or sparse Matrix of non-negative integers (the data).}

\item{quadpts}{positive integer, number of quadrature points.}

\item{maxtry}{positive integer, number of times to retry fitting if initial
fit fails due to numerical problems. Each retry increases quadpts by 500.}

\item{verbose}{logical, should progress messages be printed?}

\item{...}{additional arguments passed to \code{\link{nblomax_mle}}}
}
\value{
a data frame whose rows correspond to the columns of m.
  The columns contain the MLEs for the parameters, the log-likelihood, and
  the \code{\link[stats]{BIC}} values.
}
\description{
Compute the maximum likelihood estimates for parameters of the
negative binomial-Lomax distribution for each column of a matrix
}
