% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nblomax.R
\name{pglomax_pzero2scale}
\alias{pglomax_pzero2scale}
\title{Estimate Poisson- or geometric-Lomax scale parameter from zero
  fraction}
\usage{
pglomax_pzero2scale(lpz, lik = c("poisson", "geometric"), tail = 0.8,
  quadpts = 1000, lims = c(1e-06, 100))
}
\arguments{
\item{lpz}{a numeric vector representing the natural log of the empirical
fraction of zeros from count data.}

\item{lik}{string specifying either the Poisson or geometric likelihood.}

\item{tail}{either a positive scalar or a vector of same length as lpz
specifying the fixed tail parameter(s).}

\item{quadpts}{positive integer, number of quadrature points.}

\item{lims}{interval of possible values for the scale
parameter, passed to \code{\link[stats]{uniroot}}.}
}
\value{
a numeric vector of estimated scale parameters for each element of
  lpz.
}
\description{
Given a fixed tail parameter, use the fraction of zeros to
  estimate the scale parameter of a Poisson-Lomax or geometric-Lomax
  distribution via the method of moments.
}
