% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/poilog_user_function.R
\name{poilog_mle}
\alias{poilog_mle}
\alias{poilog_mle,Matrix-method}
\alias{poilog_mle,matrix-method}
\alias{poilog_mle,SummarizedExperiment-method}
\title{Poisson-lognormal maximum likelihood estimation}
\usage{
poilog_mle(object, ...)

\S4method{poilog_mle}{Matrix}(object, mc.cores = 1)

\S4method{poilog_mle}{matrix}(object, mc.cores = 1)

\S4method{poilog_mle}{SummarizedExperiment}(object, assayName = "counts", mc.cores = 1)
}
\arguments{
\item{object}{A \code{SingleCellExperiment}, \code{SummarizedExperiment},
\code{matrix} or sparse \code{Matrix} of UMI counts (non-negative integers)}

\item{...}{for the generic, additional arguments to pass to object-specific
methods.}

\item{mc.cores}{Positive integer indicating the number of cores to use for
parallel processing. See \code{\link[parallel]{mclapply}}.}

\item{assayName}{In case object is a SingleCellExperiment or
SummarizedExperiment, the assay containing the UMI counts.}
}
\value{
a data frame whose rows correspond to the columns of the input data.
  The columns contain the MLEs for the parameters (mu and sig),
  the log-likelihood, and
  the \code{\link[stats]{BIC}} values. NA values indicate numerical
  problems with the MLE fit.
}
\description{
Compute the maximum likelihood estimates for parameters of the
Poisson-lognormal distribution for each column of a matrix.
}
\details{
This is essentially a convenience wrapper around
  \code{\link[sads]{fitpoilog}} with optional parallelization. WARNING:
  only counts from unique molecular identifiers (UMIs) should be used to
  fit the Poisson-lognormal, as read counts and other normalized counts
  are poorly fit by this distribution.
}
