% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poilog.R
\name{poilog_pzero2mu}
\alias{poilog_pzero2mu}
\title{Estimate Poisson-lognormal scale parameter from zero fraction}
\usage{
poilog_pzero2mu(lpz, sig = 2.5, lims = c(-200, 200))
}
\arguments{
\item{lpz}{a numeric vector representing the natural log of the empirical
fraction of zeros from count data.}

\item{sig}{either a positive scalar or a vector of same length as lpz
specifying the fixed sigma parameter (see \code{\link[sads]{dpoilog}}).}

\item{lims}{interval of possible values for the scale
parameter (mu), passed to \code{\link[stats]{uniroot}}.}
}
\value{
a numeric vector of estimated scale parameters (mu) for each element
of lpz.
}
\description{
Given a fixed sigma parameter, use the fraction of zeros to
estimate the mu parameter of a Poisson-lognormal
distribution via the method of moments.
}
