% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/quminorm_user_function.R
\name{quminorm}
\alias{quminorm}
\alias{quminorm,Matrix-method}
\alias{quminorm,matrix-method}
\alias{quminorm,SummarizedExperiment-method}
\title{Quantile normalization of non-UMI single-cell gene expression}
\usage{
quminorm(object, ...)

\S4method{quminorm}{Matrix}(object, shape = 2, mc.cores = 1)

\S4method{quminorm}{matrix}(object, shape = 2, mc.cores = 1)

\S4method{quminorm}{SummarizedExperiment}(object, assayName = "tpm", shape = 2, mc.cores = 1)
}
\arguments{
\item{object}{A \code{SingleCellExperiment}, \code{SummarizedExperiment},
\code{matrix} or sparse \code{Matrix} of non-negative integers (counts).}

\item{...}{for the generic, additional arguments to pass to object-specific
methods.}

\item{shape}{Positive scalar, a fixed shape parameter for the target
distribution. The shape parameter represents sigma for the
Poisson-lognormal target distribution. See \code{\link[sads]{dpoilog}}.}

\item{mc.cores}{Positive integer indicating the number of cores to use for
parallel processing. See \code{\link[parallel]{mclapply}}.}

\item{assayName}{In case object is a SingleCellExperiment or
SummarizedExperiment, the assay from which
the quasi-UMIs should be calculated is specified here.}
}
\value{
An object of same class as m but with the nonzero values normalized
  to match the target quasi-UMI distribution. In case of a
  SingleCellExperiment or SummarizedExperiment,
  the quasiUMIs are added to a new assays slot named qumi_poilog_[shape]
  where [shape] is the user-specified numeric shape parameter.
}
\description{
Normalizes read counts (without UMIs) or
  transcripts per million (TPM) such as those generated by smart-seq2
  to match a discrete quasi-UMI target distribution. The resulting QUMI
  counts can be analyzed as if they were UMI counts.
}
\details{
The default assay of "tpm" for transcripts per million
  does refer to such values after
  summarizing to gene level. The algorithm in its current form has not been
  evaluated for quasi-UMI generation from transcript counts, as full-length
  data with UMIs have not until very recently been available.
}
\examples{
library(scRNAseq)
library(SingleCellExperiment)
sce <- ReprocessedAllenData("rsem_tpm")
sce <- quminorm(sce[,1:3],assayName="rsem_tpm")
assayNames(sce)

}
\references{
Townes FW and Irizarry RA (2020). Quantile normalization of
  single-cell RNA-seq read counts without unique molecular identifiers.
  \emph{bioRxiv}
  \url{https://doi.org/10.1101/817031}
}
