% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quminorm_ERCC_opt.R
\name{quminorm_ERCC_opt}
\alias{quminorm_ERCC_opt}
\title{ERCC-optimization of quminorm shape}
\usage{
quminorm_ERCC_opt(
  m,
  ERCC_conc_data,
  shapeVals = seq(1.4, 2.6, 0.1),
  quadpts = 1000
)
}
\arguments{
\item{m}{a matrix or sparse Matrix of non-negative integers (count data).}

\item{ERCC_conc_data}{A dataframe or matrix with two columns, where column 1
should be a character vector containing the names of the ERCC, and column 2
should contain the known concentrations in the input.}

\item{shapeVals}{The values for the shape that should be evaluated.}

\item{quadpts}{positive integer, number of quadrature points. Increase for
greater precision but slower computation.}
}
\value{
a matrix-like object of same type as m but with the nonzero values
normalized to match the target quasi-UMI distribution.
}
\description{
Optimizing the shape parameter for the quantile normalization on a per-cell
basis, by optimizing the correlation between the observed and the known
concentrations of the External RNA Controls Consortium (ERCC) standard
transcripts. The underlying assumption is that cells with different skewing
of their transcriptomes will need different shape parameters, and as the
ERCC should be uniformly PCR-skewed in all cells, they can be used to tune
this.
}
