% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quminorm.R
\name{quminorm_matrix}
\alias{quminorm_matrix}
\title{Quantile normalization of a non-UMI read counts matrix}
\usage{
quminorm_matrix(m, shape, lik = c("poilog", "plomax", "nb"),
  quadpts = 1000)
}
\arguments{
\item{m}{a matrix or sparse Matrix of non-negative integers (count data).}

\item{shape}{positive scalar, a fixed shape parameter for the target
distribution. The shape parameter represents sigma, tail, and size for the
Poisson-lognormal, Poisson-Lomax, and negative binomial target
distributions, respectively. See \code{\link[sads]{dpoilog}}, \code{\link{dnblomax}}, or
\code{\link[stats]{dnbinom}}.}

\item{lik}{likelihood of target distribution, either Poisson-lognormal,
Poisson-Lomax, or negative binomial.}

\item{quadpts}{positive integer, number of quadrature points. Increase for
greater precision but slower computation.}
}
\value{
a matrix-like object of same type as m but with the nonzero values
  normalized to match the target quasi-UMI distribution.
}
\description{
Normalizes a matrix of read counts (without UMIs) such as those
generated by smart-seq2 to match a discrete quasi-UMI target distribution.
The resulting QUMI counts can be analyzed as if they were UMI counts.
}
