% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nblomax.R
\name{rlomax}
\alias{rlomax}
\title{Lomax random number generation}
\usage{
rlomax(n, tail = 1.01, scale = 1, logscale = FALSE)
}
\arguments{
\item{n}{a positive scalar indicating the number of random variables to draw.}

\item{tail}{a positive scalar controlling the heaviness of the Lomax tail.}

\item{scale}{a positive scalar.}

\item{logscale}{logical; if TRUE, the natural log of the random variables is
returned instead of the variables themselves.}
}
\value{
a vector of (possibly log-transformed) Lomax distributed random
  variates.
}
\description{
Draw random samples from the Lomax distribution
}
\details{
The density function of the Lomax distribution is given by
\eqn{f(x)=\frac{\alpha}{\lambda}\left(1+\frac{x}{\lambda}\right)^{-(\alpha+1)}}
where \eqn{\alpha,\lambda} are the tail and scale parameters, respectively.
Lomax has a heavy power-law tail with exponent (1+tail). The smaller the
tail parameter, the heavier the tail of the distribution. If tail<=1 there
is no mean, if tail<2 there is no variance. A larger scale parameter means
fewer zeros and more large values. The scale parameter has no effect on the
power-law tail.
}
\seealso{
\code{\link{dnblomax}},\code{\link[VGAM]{rlomax}}
}
