% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poilog.R
\name{poilog_mle}
\alias{poilog_mle}
\title{Poisson-lognormal MLEs}
\usage{
poilog_mle(x, om = "BFGS", ...)
}
\arguments{
\item{x}{vector of non-negative integers (the data).}

\item{om}{optimization method to be used by \code{\link[sads]{fitpoilog}}.}

\item{...}{additional arguments passed to \code{\link[sads]{fitpoilog}}.}
}
\value{
a named numeric vector containing the MLEs of mu,sigma (see
  \code{\link[sads]{dpoilog}} for details).
  The maximized value of the log-likelihood is attached as an attribute.
}
\description{
Compute the maximum likelihood estimates for parameters of the
  Poisson-lognormal distribution.
}
